third_party/blink/web_tests/fast/media/viewport-media-query-synchronous-expected.txt

Test for bug 75633

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(document.getElementById("test")).backgroundColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE

foo
